local_server/requirements.txt
aio-pika
pika
psycopg2-binary
sqlalchemy
sqlacodegen
pydantic
cryptography
requests
deepface
pillow
pyotp
pyzbar
easyocr
tf-keras
ultralytics
opencv-contrib-python